move everything important to server, give in to  temptation
