+++ Content.Client/Access/IdCardSystem.cs
@@ -1,0 +1,7 @@
+﻿using Content.Shared.Access.Systems;
+
+namespace Content.Client.Access;
+
+public sealed class IdCardSystem : SharedIdCardSystem
+{
+}
